Split out types into its own file
